+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- `createStream()` method on `Service` and `Messenger` to accept `Record<string | number, any>` metadata instead of `Record<any, any>`.
+
+### Added
+
+- More info about `metaData` to **Streaming data between threads** section of README.
+
 ## [0.4.3] 2023-2-3
 
 ### Changed