+++ CHANGELOG.md
@@ -18,7 +18,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 - `entries()` generator method on `SharedMap` to be more performant.
-- `Messenger.closeAll()` not closing the current messenger instance as well.
+- `Messenger.closeAll()` not closing the current messenger instance.
+- Improved general performance for services.
+-
 
 ## [0.4.3] 2023-2-3