+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `entries()` generator method on `SharedMap` to be more performant.
 - `Messenger.closeAll()` not closing the current messenger instance.
 - Improved general performance for services.
+- Improved general performance for task workers & services.
 
 ## [0.4.3] 2023-2-3