+++ CHANGELOG.md
@@ -15,7 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - More info about `metaData` to **Streaming data between threads** section of README.
 - `Messenger.closed` getter.
 - `Service.setRef()` method.
-- `shareEnv` option for task & service workers, defaulting to `true`.
+- `shareEnv` option for task & service workers, defaulting to `true` + docs.
 
 ### Fixed
 
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `Messenger.closeAll()` not closing the current messenger instance.
 - Improved general performance for services.
 - Improved general performance for task workers & services.
+- Incorrect docs for `reffed` config option. Defaults to `true`, not `false`.
 
 ## [0.4.3] 2023-2-3