+++ CHANGELOG.md
@@ -8,6 +8,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- `createStream()` method on `Service`, `Messenger`, and `ParentThread` to accept `Record<string | number, any>` metadata instead of `Record<any, any>`.
+
+### Added
+
+- More info about `metaData` to **Streaming data between threads** section of README.
+- `Messenger.closed` getter.
+- `Service.setRef()` method.
+- `shareEnv` option for task & service workers, defaulting to `true` + docs.
+
+### Fixed
+
+- `entries()` generator method on `SharedMap` to be more performant.
+- `Messenger.closeAll()` not closing the current messenger instance.
+- Improved general performance for services.
+- Improved general performance for task workers & services.
+- Incorrect docs for `reffed` config option. Defaults to `true`, not `false`.
+
+## [0.4.3] 2023-2-3
+
+### Changed
+
 - Some minor refactors.
 
 ### Added